fix: add set -e and try to mount differently
The init script of the initramfs is failing with exit code -13. No clue
what that is. Thus I'm trying to solve it at random and figure it out
afterwards. I have add the "set -e" to the init script to see if that
prints out anything. And I have also moved the clear command to run
after the "dmesg -n 1" to ensure that isn't causing the problem. If the
clear command works it's certain that the dmesg is causing the error.